streamlit/requirements.txt
streamlit
Pillow
numpy
streamlit
tensorflow-cpu==2.6.2
regex
nltk